btf: move kernel spec parsing to separate file
No changes to the source code.
